uldrenite: Remove unused motionsense node
The alternate lid accel was never used since it was missing a ssfc
flag to enable it. The only difference between this node and the lid
accel was the on-body detection. But uldrenite uses the base accel for
on-body detection.

BUG=b:410646158
TEST=flash uldrenite and uldrenite-ish check that on-body log messages
     are seen at the appropriate times.

Change-Id: Iabc9e2c5674a9f780de1c05a2fc16d26fc51b71c

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6505330
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: caveh jalali <caveh@chromium.org>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>