test/run_device_tests: Update expected power for dragonclaw v4 boards
The numbers for MCU are higher and the numbers for the sensor are lower.
I adjusted these values based on ToT measurements, but it doesn't appear
to be a regression. I also get higher numbers for the MCU measurement
when using the same version of the firmware that we measured with
dragonclaw v3: https://crrev.com/c/6527405.

BUG=b:358679133
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       -t power_utilization_sleep
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       -t power_utilization_idle
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       --zephyr -t power_utilization_sleep
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       --zephyr -t power_utilization_idle

Change-Id: Iadfdf60cbac51708ced04a4976c76e63ccd77b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6527406
Tested-by: Tom Hughes <tomhughes@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>