pdc: Support flashing RTK PDCs in ROM mode using console PDC updater
Update the EC console-based RTK PDC updater to support flashing the PDC
if it is in ROM mode. This is detected based on the GET_IC_STATUS
response. When the PDC is in ROM mode, perform an erase of flash before
flashing the new image, and skip the validation step as it is not
supported.

BUG=b:417489560
TEST=Erase a PDC, and reflash it.

Cq-Depend: chromium:6544873
Change-Id: I5a75fb6b257036b71eb18359eea708264abff7b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6542451
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>