body_detect: Notify host only in S0
When the DUT is in a suspended state S0ix we don't need to notify the
host of the change as the thermals are not a concern in suspend. This
allows the DUT to remain suspended longer and saves power.

To achieve this we need to expose some motion_sense values/functions
that were prioviously only exposed for tests.

BUG=b:410646158
TEST=flash uldrenite, run suspend_stress_test -c 10
TEST=flash uldrenite, powerd_dbus_suspend, rotate screen past 180,
     check that it woke up the DUT

Cq-Depend: chromium:6558033
Change-Id: Iab654795791402422e2832a6491efe0da43af7f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6557205
Tested-by: Yuval Peress <peress@google.com>
Auto-Submit: Yuval Peress <peress@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>