body_detection: Support ISH based body detection
Allow the ISH to notify the EC of body detection events. Similarly,
when notified, the EC will notify the AP. The AP notification is
enabled by removing the requirements !body_detect_is_remote(). This
check was previously added to avoid the EC notifying the AP as it was
believed to be possible for the ISH to signal this change directly.
Intel confirmed that this is not possible and that the notification
MUST go through the EC.

BUG=b:414696917
TEST=Run on Uldrenite, see bug for details.

Change-Id: Ie4e90858af53212804516ba7016fb864f0f0de83

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6616939
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Auto-Submit: Yuval Peress <peress@google.com>
Code-Coverage: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>