zephyr: hwtest: panic_data: Add workaround for LTO
When LTO is enabled the link register does not match the expected value.
The issue is that the __ASSERT_NO_MSG is outlined into its own function,
so the link register points to the address of the outlined function
instead of crash_system:

llvm-objdump -d build/zephyr/bloonchipper/build-rw/zephyr/zephyr.elf

08089c56 <OUTLINED_FUNCTION_5>:
 8089c56: f84d ed08     str     lr, [sp, #-8]!
 8089c5a: 4620          mov     r0, r4
 8089c5c: 217e          movs    r1, #0x7e
 8089c5e: f7f8 fedf     bl      0x8082a20 <assert_post_action> @ imm = #-0x7242
 8089c62: f85d eb08     ldr     lr, [sp], #8
 8089c66: defe          trap    # <--- This is the assert

If we use __ASSERT_UNREACHABLE (the last part of the __ASSERT_NO_MSG
macro), then it won't be outlined, but it is inlined into test_crash.

BUG=b:423904871, b:407119021
TEST=./test/run_device_tests.py --zephyr -t panic_data

Change-Id: I682744c1d1c801705b6de8e6752a619bcb2f219f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6635116
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>