pdc: Allow comms suspension if stuck in init state w/ error
Normally, suspending is not allowed while pdc_power_mgmt is in its init
state so that it can complete its initialization steps cleanly before
suspending. However, if there is a serious error with the PDC, allow the
pdc_power_mgmt state machine to suspend to allow a PDC FW update to
proceed. This is useful for PDC testing when a PDC might be in ROM code
or the devicetree is misconfigured during bringup.

BUG=b:417489560
TEST=./twister -s pdc/pdc.generic
TEST=Live test with a blank PDC

Change-Id: Ib5736fb9dba364456b774063c06fe4b1bc760f71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6649297
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>