pdc: tps6699x: Don't send READ_POWER_LEVEL as sink
The PDC power manager uses the SNK_ATTACHED_READ_POWER_LEVEL state
because the RTK PDC relies on this to update the ADC read. But, this is
an invalid UCSI command. The UCSI 3.0 spec states the following.

"This command is used by the OPM to read a port’s peak power and average
power levels from the LPM that is in the sourcing mode. If the connector
does not have an active connection or the LPM is in the sink mode, this
command has no effect, and the PPM shall set the Error Information field
to indicate Invalid Command Specific Parameters."

The command always fails for the TI PDC and it contributes to a
TEST.PD.PS.SNK.1 compliance failure. This CL updates the TI PDC driver
to return -EINVAL if the PDC power manager attempts sending this as a
Sink.

BUG=b:424219903
TEST=TEST.PD.PS.SNK.1 passing

Change-Id: I6d4d7c89517e430f5c7ddc86d940447385aec671
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6653567
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Tested-by: Jameson Thies <jthies@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>