pdc: Config charge manager seed to 0V/0A if disabling a port
Due to changes to the initial seeding of the charge manager if a port is left
uninitialized the charge manager will not allow the system to boot without a
battery. In the case of a port is being disabled at a software level init the
charge manager to 0V/0A to avoid having uninitialized values.

BUG=b:423808174
TEST=zmake build ptlrvp_mchp, verified previously not booting board boots
LOW_COVERAGE_REASON=Coverage to be handled in b:427712128

Change-Id: Idd0346b0d255423c62ab7dc1863bebdd26311813
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6655359
Commit-Queue: caveh jalali <caveh@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Brandon Breitenstein <brandon.breitenstein@intel.com>
Reviewed-by: caveh jalali <caveh@chromium.org>