skywalker: add BUILD_ASSERT for hardcoded dp_aux_path_sel index
This patch adds a compile-time assert to ensure the consistency of the
`dp_aux_path_sel` GPIO index.

The `dp_aux_path_sel` index is hardcoded in the kernel (CL:6654443) and
must not change after FSI. This assertion acts as a safeguard, triggering
a compile-time error if the index is changed.

BUG=b:418873477
TEST=zmake build --all

Change-Id: I9695801cb5f219e13473d671bb377c391f4dba3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6656400
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>