lid_angle: Reset to clamshell on remote lid-angle reboot
When the ISH reboots it resets the state and assumes it's starting up
in clamshell mode until it calculates the lid angle. The issue came
up when the DUT was placed in clamshell while the DUT was off. This
left the EC thinking it was in tablet mode while the new ISH instance
assumed (correctly) that it was in clamshell. Since the ISH didn't
detect a change, it never notified the EC. Instead of adding another
communication pass during initialization, the EC should just reset its
assumption on chip startup.

BUG=b:427631462
TEST=Enter tablet mode,
     Shut down,
     Power on -> tablet mode
TEST=Enter tablet mode,
     Shut down,
     Enter clamshell mode,
     Power on -> clamshell mode
TEST=Enter clamshell mode,
     Shut down,
     Power on -> clamshell mode
TEST=Enter clamshell mode,
     Shut down
     Enter tablet mode,
     Power on -> tablet mode

Change-Id: I5c15d3965632f7980748623662f822b37d9a5bf4

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6682257
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Auto-Submit: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Allan Almiranez <allanalmiranez@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Code-Coverage: Yuval Peress <peress@google.com>