motion_sense: Make activity sensor ID unsigned always
The ternary for `? motion_sensor_count : -1` was causing an error with
strict checks since the motion_sensor_count is unsigned. Instead of
relying on the compiler to cast -1 to a uint8_t, we can create an
invalid ID at 0xff and check against that.

BUG=b:281981515,b:424657530
TEST=CQ

Change-Id: I9a20c0971fa6e8a00bf49b4fb072a508902ea109

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6689095
Commit-Queue: Yuval Peress <peress@google.com>
Code-Coverage: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>