pdc: ti: Copy PDO values in tps_set_pdos
`tps_set_pdos()` holds on to a caller-provided pointer to an array of
PDOs to set. This can lead to a memory access bug because the command is
queued up and might not get executed before the caller returns. Make a
copy of the PDO array before returning.

BUG=None
TEST=./twister -T zephyr/test/pdc

Change-Id: I3efdd01d9916b6171c05fcf007ffc4bb4789b42e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6701765
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>