col_gpio_drive: Add BUILD_ASSERT for SSPI/KSO16/17 pin conflict
On ITE ECs, the hardware pins for the spi0 controller's SMOSI and
SMISO signals are multiplexed with the keyboard matrix output pins
KSO16 and KSO17. If spi0 is enabled in the devicetree, the keyboard
controller must be configured to not drive KSO16 and KSO17 to
prevent a hardware pin conflict.

This change adds a BUILD_ASSERT to enforce this constraint at
compile time.

BUG=b:435516562
TEST=Passed that correctly config kso_ignore_mask when spi0 is enabled.
TEST=Failed that spi0 is enabled but kso_ignore_mask is missing bits.
TEST=Passed that spi0 is disabled, regardless of kso_ignore_mask.
TEST=zmake build --all

Change-Id: I3468e5ee6e774dfc2380e633162d83a2697761cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6829048
Tested-by: Wai-Hong Tam <waihong@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>