fpsensor: Remove template unlock mechanism
According to go/eliminate-fpmcu-preload the template lock/unlock
mechanism is used to make sure that the rate limiter in GSC is not
bypassed. The match can be performed only once in a nonce context.
To perform another match, we need to get nonce context again from GSC
and unlock templates (extra step to make sure that set of templates
is associated with nonce context),

In the new Fingerprint Auth approach the Trusted Application will send
match request signed with the session key and challenge that must be
signed by the FPMCU when the match is successful. The new approach does
not require reloading nonce context (which will be eventually
deprecated), so template locking/unlocking is not necessary.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands

Change-Id: I313104f8ec8bb4871af1b0d76dda31a2df697b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6839790
Commit-Queue: Patryk Duda <patrykd@google.com>
Code-Coverage: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>