emul: tps6699x: Correct handling of CCOM value
The TI PDC emulator decodes the bitfield from the SET_CCOM command and
stores a corresponding enum value for the CC operating mode (CCOM).
However, `tps699x_emul_set_pdr()` assumes the value stored in
`data->ccom` is the bitfield. Additionally, the emualtor reset function
loads BIT(2) as the default value for `data->ccom`, but it should be the
CCOM_DRP enum, which and has a value of 2, not BIT(2).

The SET_PDR handler checks the CCOM mode to determine how it reacts to
power role change requests sent through SET_PDR. Because it only does so
when the CCOM is BIT(2) (instead of the intended CCOM_DRP), any change
to the CCOM, even setting it to DRP, causes the emualtor to remain in
its current power role when calling SET_PDR.

Update `tps699x_emul_set_pdr()` to check for CCOM_DRP and not BIT(2).
Initialize the field to CCOM_DRP instead of BIT(2). Also add more useful
logging to the SET_PDR and SET_CCOM handlers.

BUG=b:435495591
TEST=./twister -T zephyr/test/pdc

Change-Id: I5710f2e0e26095aeeebd3e5181f2c4a3eb348a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6842143
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Tristan Honscheid <honscheid@google.com>