pdc: Add logs for should_swap_to_source()
This function is called to evaluate if we should stop sinking and become
a source, but has no logging currently. This can lead to confusing
abrupt changes to the power role without explanation.

Add logs to communicate that the check is being run and explain the
reason for not swapping to source

BUG=b:432292811
TEST=Flash to lotso device
LOW_COVERAGE_REASON=Log-only change. b/339686045

Change-Id: I33c5391667a3ef2a35d6e8f7abc2ee624f14d314
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6847600
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>