brox: Fix Host_Speed in Brox VIF
Both ports' host speed on SKU1 is 2x2, but the VIF incorrectly listed 1x2.

This came up during informational Link Layer testing at USB Compliance
Workshop #138. With the VIF as presented, we could not run Gen 2 tests
because we indicated we only supported Gen 1.

With this changed to more accurately list 2x2, the Teledyne LeCroy m310e
was able to run Gen 2x1 tests on Brox.

Also, with this change, this VIF now only refers to the behavior of SKU1,
as SKU2 (with a Burnside Bridge retimer) has a diffent maximum host speed
because of the limitation imposed by the BB retimer. Rename the vif to
make clear this is a specific vif for SKU1, and so future folks trying to
test with SKU2 need to create a separate VIF.

BUG=b:439665648
TEST=Run Link Layer testing for Gen 2 on Teledyne Lecroy m310e. Confirm
test runs and isn't completely skipped.

Change-Id: I1b26a85c5074d4be5e7b7eb4086e388dbe36d295
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6863716
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Benson Leung <bleung@google.com>
Commit-Queue: Benson Leung <bleung@google.com>
Code-Coverage: Benson Leung <bleung@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>