fpsensor: Add template confirm host command
In the Fingerprint Auth approach, the successfully enrolled template
must be confirmed, in order to be downloaded and used for matching.

This is necessary, because AuthenticatorID in the Trusted Application,
must be changed only after new template was enrolled. If the enroll
session was canceled by the user the AuthenticatorID must remain the
same.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands

Change-Id: I8d698cb8ecf620f1ec3adc4cd99c8f2216136b26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6924454
Code-Coverage: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>